requirements.txt
# Core dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
gitpython>=3.1.0
PyGithub>=2.1.0

# LLM
openai>=1.0.0
anthropic>=0.18.0  # opsiyonel

# CLI
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Parsing
# tree-sitter>=0.20.0  # opsiyonel, multi-language için

# Utils
pyyaml>=6.0
chardet>=5.0.0
tqdm>=4.65.0
tiktoken>=0.5.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
